+++ Sources/SuperwallKit/Network/Device Helper/DeviceHelper.swift
@@ -200,33 +200,53 @@ class DeviceHelper {
     if #available(iOS 14.5, *) {
       networks += [.mir]
     }
+
     if #available(iOS 15.0, *) {
       networks += [.nanaco]
     }
     if #available(iOS 15.1, *) {
       networks += [.dankort]
     }
+
+    #if compiler(>=5.7)
     if #available(iOS 16.0, *) {
       networks += [.bancomat, .nanaco, .waon]
     }
+    #endif
+
+    #if compiler(>=5.8)
     if #available(iOS 16.4, *) {
       networks += [.postFinance]
     }
+    #endif
+
+    #if compiler(>=5.9)
     if #available(iOS 17.0, *) {
       networks += [.tmoney, .pagoBancomat]
     }
-    if #available(iOS 17.4, *) {
+    #endif
+
+    #if compiler(>=5.10)
+    if #available(iOS 17.4, visionOS 1.1, *) {
       networks += [.meeza]
     }
-    if #available(iOS 17.5, *) {
+
+    if #available(iOS 17.5, visionOS 1.2, *) {
       networks += [.bankAxept, .NAPAS]
     }
-    if #available(iOS 18.4, *) {
+    #endif
+
+    #if compiler(>=6.1)
+    if #available(iOS 18.4, visionOS 2.4, *) {
       networks += [.himyan, .jaywan]
     }
-    if #available(iOS 26.0, *) {
+    #endif
+
+    #if compiler(>=6.2)
+    if #available(iOS 26.0, visionOS 26.0, *) {
       networks += [.myDebit]
     }
+    #endif
 
     return PKPaymentAuthorizationViewController.canMakePayments(
       usingNetworks: networks