+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.10.6
+
+### Fixes
+
+- Fixes issue that prevented the SDK from being built on old Xcode versions.
+
 ## 4.10.5
 
 ### Fixes
 
 - Updates `device.isApplePayAvailable` for more accurate filtering. Previously it returned true whenever the device supported Apple Pay, even if no card was added. It now returns true only when the device supports Apple Pay and the user has added a card.
+- Fixes issue where `didRedeemLink` might not get called if there's no paywall available to present an alert from.
 
 ## 4.10.4